delete syllabus added

+++ syllabusRoute/deleteSyllabus.js
@@ -2,6 +2,7 @@ let db = require('./databaseQuerySyllabus')
 let errorCode = require('../commonFunction/errorCode')
 let getCode = new errorCode()
 let deleteSyllabus;
+let isInSchool;
 let id;
 
 module.exports = require('express').Router().post('/',async(req,res)=>{
@@ -14,6 +15,14 @@ module.exports = require('express').Router().post('/',async(req,res)=>{
                 "message" : "Id not given",
                 status_name : getCode.getStatus(404)
             })
+        }
+        isInSchool = await db.selectSchool(id)
+        if(isInSchool.length > 0){
+            return res.json({
+                "status_code" : 1063,
+                "message" : "Syllabus alreay linked with some schools",
+                status_name : getCode.getStatus(1063)
+            }) 
         }
         deleteSyllabus = await db.deleteSyllabus(id);
         if(deleteSyllabus.affectedRows > 0){
